+++ CHANGELOG.md
@@ -3,11 +3,17 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [1.0.1] - 2025-06-15
 
 ### Fixed
 
-- Updated IP Info widget with a more reliable service and more frequent updates
+- Switched to /thumbnail for viewing non-web friendly image formats (ex: HEIC)
+
+### Added
+
+- Automatic album refresh once a week
+
+# PRE-FORK CHANGELOG
 
 ## [2.6.0] - 2022-04-25